+++ R/adjust_mortality.R
@@ -1,16 +1,15 @@
 #' Adjust interpolated tables for mortality
 #'
-#' Trees in the input `data_interpolated` already have had their switch to
-#' `STATUSCD` 2 (i.e. death) interpolated to the midpoint (rounded down) between
-#' it's last survey alive and first survey dead.
+#' This is an "internal" function—most users will want to run [fia_annualize()]
+#' instead. Trees in the input `data_interpolated` already have had their switch
+#' to `STATUSCD` 2 (i.e. death) interpolated to the midpoint (rounded down)
+#' between it's last survey alive and first survey dead.
 #'
 #' This does the following:
 #' - Optionally figures out if a tree has a recorded `MORTYR` and uses that for
 #'   the transition to `STATUSCD` 2 instead of the interpolated values. If the
 #'   tree is alive (`STATUSCD` 1) in `MORTYR`, then it is assumed it died in the
 #'   following year.
-#' - Drops trees that transition to `STATUSCD` 0 and `RECONCILECD` 5, 6, or 9
-#'   (moved out of plot) at the midpoint between surveys
 #' - Adjusts `STANDING_DEAD_CD` so that it only applies to dead trees
 #' - Adjusts `DECAYCD` so that it only applies to standing dead trees
 #' - Adjusts `DIA`, `HT`, `ACTUALHT`, `CULL`, and `CR` so that they only apply
@@ -20,6 +19,7 @@
 #' @param use_mortyr logical; use `MORTYR` (if recorded) as the first year a
 #'   tree was dead?
 #' @export
+#' @keywords internal
 #' @returns a tibble
 adjust_mortality <- function(data_interpolated, use_mortyr = TRUE) {
   cli::cli_progress_step("Adjusting for mortality")
@@ -45,7 +45,7 @@ adjust_mortality <- function(data_interpolated, use_mortyr = TRUE) {
     # If a tree is marked as alive (STATUSCD 1) in the recorded MORTYR (i.e.
     # MORTYR is an inventory year), then assume death happened the year after
     # that inventory year
-    # (https://github.com/mekevans/forestTIME-builder/issues/61)
+    # (https://github.com/Evans-Ecology-Lab/forestTIME-builder/issues/61)
 
     df <- data_interpolated |>
       dplyr::group_by(tree_ID) |>
@@ -59,7 +59,7 @@ adjust_mortality <- function(data_interpolated, use_mortyr = TRUE) {
       ) |>
       # this "fills in" the new effective MORTYR for all rows since the above
       # if_else() only increments it for the inventory year row
-      mutate(MORTYR_eff = max(MORTYR_eff)) |>
+      dplyr::mutate(MORTYR_eff = max(MORTYR_eff)) |>
       dplyr::mutate(
         # STATUSCD is interpolated to the midpoint between surveys
         # see utils.R for more info on what %|||% does
@@ -67,7 +67,7 @@ adjust_mortality <- function(data_interpolated, use_mortyr = TRUE) {
       ) |>
       # When a tree has a recorded MORTYR, adjust STATUSCD depending on whether
       # MORTYR is before or after the midpoint (first_dead). This works because
-      # MORTYR is filled in for every row of a tree by prep_data() and
+      # MORTYR is filled in for every row of a tree by fia_tidy() and
       # expand_data(). Can't assume tree has STATUSCD 2 after MORTYR since
       # sometimes STATUSCD goes from 1 to 2 to 0.
       dplyr::mutate(
@@ -93,7 +93,40 @@ adjust_mortality <- function(data_interpolated, use_mortyr = TRUE) {
     df <- data_interpolated |> dplyr::group_by(tree_ID)
   }
 
-  df |>
+# If trees are interpolated to below FIA thresholds for being measured, set
+  # them to fallen dead. For most trees, this is DIA < 1 and ACTUALHT < 4.5.
+  # For woodland species, the ACTUALHT threshold is 1. To figure out if a tree
+  # is a woodland species, we need to pull in one of the ref tables
+  # temporarily.
+
+  ref_species <-
+    REF_SPECIES |>
+    dplyr::select(
+      SPCD,
+      JENKINS_SPGRPCD
+    )
+
+  data_adjusted <- df |>
+    dplyr::left_join(ref_species, by = dplyr::join_by(SPCD)) |> 
+    # TODO: Is there a way of only having to do the case_when once?  E.g. would
+    # it be faster to create a column "dead_fallen" and then in a subsequent
+    # step use dead_fallen to set STATUSCD and STANDING_DEAD_CD? this function
+    # feels a lot slower since adding this bit
+    dplyr::mutate(
+      STATUSCD = dplyr::case_when(
+        JENKINS_SPGRPCD < 10 & (DIA < 1 | HT < 4.5 | ACTUALHT < 4.5) ~ 2,
+        JENKINS_SPGRPCD == 10 & (DIA < 1 | HT < 1 | ACTUALHT < 1) ~ 2,
+        .default = STATUSCD
+      ),
+      STANDING_DEAD_CD = dplyr::case_when(
+        JENKINS_SPGRPCD < 10 & (DIA < 1 | HT < 4.5 | ACTUALHT < 4.5) ~ 0,
+        JENKINS_SPGRPCD == 10 & (DIA < 1 | HT < 1 | ACTUALHT < 1) ~ 0,
+        .default = STANDING_DEAD_CD
+      )
+    ) |> 
+    dplyr::select(-JENKINS_SPGRPCD)
+
+  data_adjusted |>
     # adjust STATUSCD & DECAYCD
 
     # STANDING_DEAD_CD only applies to dead trees
@@ -112,7 +145,7 @@ adjust_mortality <- function(data_interpolated, use_mortyr = TRUE) {
       )
     ) |>
     # trees in non-sampled areas shoudn't have measurements for anything
-    # https://github.com/mekevans/forestTIME-builder/issues/59
+    # https://github.com/Evans-Ecology-Lab/forestTIME-builder/issues/59
     dplyr::mutate(
       dplyr::across(
         c(DIA, HT, ACTUALHT, CULL, CR),