+++ R/data.R
@@ -1,5 +1,5 @@
 #' State codes used in FIA
 #'
-#' A data frame tha matches the `STATECD`` column used in FIA data to the two
-#' letter state abbreviations.
-"state_codes"
+#' A data frame tha matches the `STATECD` column used in FIA data to the two
+#' letter state abbreviations and the total land area of the state in acres.
+"state_areas"